Remove debug and release source sets
They were previously created to reduce the number of calls to the server during the development. But it is not convenient to manage source sets for this reason only. A simpler approach would be to simply configure the app to refetch the grocery database only on release. (The initial prepopulation will still fetch the data from the server, but only once).